Tweak, move `HallSymbol` to separate file
