+++ hikari/symmetry/__init__.py
@@ -5,7 +5,7 @@ to be used by other objects.
 import pickle
 from .operations import SymmOp
 from .group import Group
-from .space_groups import HallSymbol
+from .hall_symbols import HallSymbol
 from hikari.resources import point_groups_pickle
 from hikari.resources import space_groups_pickle
 PG = pickle.loads(point_groups_pickle)
